issue on clean task
Hi,
There is an issue with the clean task. it should be like this:
del(['/dist'], cb);

The "/" is missing so it will throw an error.

I want to tell you that this is a awesome generator, thank you and keep going on it.
You save me a lot of time.